Add support for Glance store configuration
As of today, Glance is configured with the filesystem store unless the
swift screen session is started in which case the swift store will be
used.
In this patch, we provide the ability to run stack.sh with a specific
store and provide implementation for the VMware store and Swift.
Support for other stores (such as S3) will come in subsequent patches.

The localrc configuration for the VMware store should be:
GLANCE_STORE=vsphere
VMWAREAPI_IP=vcenter_ip
VMWAREAPI_USER=vcenter_login
VMWAREAPI_PASSWORD=vcenter_pw
VMWAREAPI_DATACENTER=datacenter_path
VMWAREAPI_DATASTORE=datastore_name

Change-Id: I3a0da29a9ded5708805e5971983f4ed75763ae93